rival_project/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Mail
email-validator
marshmallow
requests
beautifulsoup4
pandas
numpy
gunicorn
python-dotenv
yfinance
psycopg[binary]